requirements.txt
# Core dependencies
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
python-dotenv>=1.0.1
requests>=2.31.0

# AI/ML dependencies
crewai>=0.28.0
google-generativeai>=0.3.2
groq>=0.4.2

# Additional utilities
typing-extensions>=4.9.0
python-multipart>=0.0.9 